+ marking/switching to current kitten on table/scroll segue * signed/unsigned -> NSInteger

+++ CatShop/KittenScrollViewController.h
@@ -7,6 +7,7 @@
 
 @property (weak) IBOutlet UIScrollView *scrollViewOutlet;
 
-@property unsigned cacheNextViewsAmount;
+// cacheNextViewsAmount must be positive
+@property NSInteger cacheNextViewsAmount;
 
 @end
